Implement changes from 4/6/22 code review with Lee
