+++ desktop/cmp/dash/canvas/DashCanvas.js
@@ -6,17 +6,17 @@ import PT from 'prop-types';
 
 import ReactGridLayout, {WidthProvider} from 'react-grid-layout';
 import 'react-grid-layout/css/styles.css';
-import './DashReport.scss';
-import {DashReportModel} from './DashReportModel';
-import {dashReportContextMenu} from './impl/DashReportContextMenu';
-import {dashReportView} from './impl/DashReportView';
+import './DashCanvas.scss';
+import {DashCanvasModel} from './DashCanvasModel';
+import {dashCanvasContextMenu} from './impl/DashCanvasContextMenu';
+import {dashCanvasView} from './impl/DashCanvasView';
 
 const reactGridLayout = elemFactory(WidthProvider(ReactGridLayout));
 
-export const [DashReport, dashReport] = hoistCmp.withFactory({
-    displayName: 'DashReport',
-    className: 'xh-dash-report',
-    model: uses(DashReportModel),
+export const [DashCanvas, dashCanvas] = hoistCmp.withFactory({
+    displayName: 'DashCanvas',
+    className: 'xh-dash-Canvas',
+    model: uses(DashCanvasModel),
     render({className, model}) {
         return div({
             className: classNames(
@@ -28,8 +28,8 @@ export const [DashReport, dashReport] = hoistCmp.withFactory({
             onContextMenu: (e) => {
                 const {clientX: x, clientY: y} = e;
                 ContextMenu.show(
-                    dashReportContextMenu({
-                        dashReportModel: model,
+                    dashCanvasContextMenu({
+                        dashCanvasModel: model,
                         clickPosition: {x, y}
                     }),
                     {left: x, top: y},
@@ -41,8 +41,8 @@ export const [DashReport, dashReport] = hoistCmp.withFactory({
                 layout: model.layout,
                 cols: model.columns,
                 rowHeight: model.rowHeight,
-                isDraggable: model.isDraggable,
-                isResizable: model.isResizable,
+                isDraggable: !model.layoutLocked,
+                isResizable: !model.layoutLocked,
                 compactType: model.compact ? 'vertical' : null,
                 margin: model.margin,
                 maxRows: model.maxRows,
@@ -55,13 +55,13 @@ export const [DashReport, dashReport] = hoistCmp.withFactory({
                 onLayoutChange: (layout) => model.setLayout(layout),
                 items: model.viewModels.map(vm => div({
                     key: vm.id,
-                    item: dashReportView({model: vm})
+                    item: dashCanvasView({model: vm})
                 }))
             })
         });
     }
 });
 
-DashReport.propTypes = {
-    model: PT.oneOfType([PT.instanceOf(DashReportModel), PT.object])
+DashCanvas.propTypes = {
+    model: PT.oneOfType([PT.instanceOf(DashCanvasModel), PT.object])
 };
