+++ desktop/cmp/dash/canvas/DashCanvas.scss
@@ -1,4 +1,4 @@
-.xh-dash-report {
+.xh-dash-Canvas {
   width: 100%;
   height: 100%;
   overflow: auto;
