+++ desktop/cmp/dash/canvas/impl/DashCanvasContextMenu.js
@@ -11,19 +11,19 @@ import {Icon} from '@xh/hoist/icon';
 import {isEmpty} from 'lodash';
 
 /**
- * Context menu to add items to a DashReport
+ * Context menu to add items to a DashCanvas
  *
  * Available view specs are listed in their defined order, optionally
  * grouped by their 'groupName' property
  *
- * @see DashReportModel
+ * @see DashCanvasModel
  * @private
  */
-export const dashReportContextMenu = hoistCmp.factory({
+export const dashCanvasContextMenu = hoistCmp.factory({
     model: null,
     observer: null,
-    render({dashReportModel, clickPosition}) {
-        const menuItems = createMenuItems({dashReportModel, clickPosition});
+    render({dashCanvasModel, clickPosition}) {
+        const menuItems = createMenuItems({dashCanvasModel, clickPosition});
         return contextMenu({menuItems});
     }
 });
@@ -31,14 +31,14 @@ export const dashReportContextMenu = hoistCmp.factory({
 //---------------------------
 // Implementation
 //---------------------------
-function createMenuItems({dashReportModel, clickPosition}) {
-    const addMenuItems = createAddMenuItems({dashReportModel, clickPosition}),
-        {extraMenuItems} = dashReportModel;
+function createMenuItems({dashCanvasModel, clickPosition}) {
+    const addMenuItems = createAddMenuItems({dashCanvasModel, clickPosition}),
+        {extraMenuItems, contentLocked} = dashCanvasModel;
     return [
         {
             text: 'Add',
             icon: Icon.add(),
-            disabled: isEmpty(addMenuItems),
+            disabled: contentLocked || isEmpty(addMenuItems),
             items: addMenuItems
         },
         ...(extraMenuItems ? ['-', ...extraMenuItems] : [])
@@ -46,13 +46,11 @@ function createMenuItems({dashReportModel, clickPosition}) {
     ];
 }
 
-function createAddMenuItems({dashReportModel, clickPosition}) {
-
-
+function createAddMenuItems({dashCanvasModel, clickPosition}) {
     const groupedItems = {},
         ungroupedItems = [],
         {x, y} = clickPosition,
-        addPosition = calcAddPosition(x, y, dashReportModel);
+        addPosition = calcAddPosition(x, y, dashCanvasModel);
 
     const addToGroup = (item, groupName) => {
         const group = groupedItems[groupName];
@@ -63,12 +61,12 @@ function createAddMenuItems({dashReportModel, clickPosition}) {
         }
     };
 
-    dashReportModel.viewSpecs
+    dashCanvasModel.viewSpecs
         .filter(viewSpec => {
             return viewSpec.allowAdd &&
                 (
                     !viewSpec.unique ||
-                    !(dashReportModel.getItemsBySpecId(viewSpec.id).length)
+                    !(dashCanvasModel.getItemsBySpecId(viewSpec.id).length)
                 );
         })
         .forEach(viewSpec => {
@@ -76,7 +74,7 @@ function createAddMenuItems({dashReportModel, clickPosition}) {
                 item = {
                     text: title,
                     icon: icon,
-                    actionFn: () => dashReportModel.addView(
+                    actionFn: () => dashCanvasModel.addView(
                         id,
                         addPosition
                     )
@@ -112,9 +110,9 @@ function createAddMenuItems({dashReportModel, clickPosition}) {
  * context menu is triggered
  * @param {number} x - clientX position
  * @param {number} y - clientY position
- * @param {DashReportModel}
+ * @param {DashCanvasModel}
  */
-const calcAddPosition = (x, y, dashReportModel) => {
+const calcAddPosition = (x, y, dashCanvasModel) => {
     const calcXY = (positionParams, top, left, w=0, h=0) => {
         const calcGridColWidth = (positionParams) => {
             const { margin, containerPadding, containerWidth, cols } = positionParams;
@@ -136,7 +134,7 @@ const calcAddPosition = (x, y, dashReportModel) => {
         return { x, y };
     };
 
-    const {margin, columns: cols, rowHeight, maxRows, ref, containerPadding} = dashReportModel,
+    const {margin, columns: cols, rowHeight, maxRows, ref, containerPadding} = dashCanvasModel,
         containerPosition = ref.current.getBoundingClientRect(),
         {left: containerLeft, top: containerTop, width: containerWidth} = containerPosition,
         positionParams = {margin, cols, rowHeight, maxRows,
